requirements.txt
streamlit
pandas
numpy
plotly
sqlalchemy
psycopg2-binary
python-dotenv
sentence-transformers
faiss-cpu
openai
